+++ projects/demo-maps/src/app/route-components/route-example-cesium/route-example-cesium.component.ts
@@ -5,7 +5,6 @@ import { MapStateService, IMapState } from '@dlr-eoc/services-map-state';
 import { OsmTileLayer, EocLitemap, BlueMarbleTile, EocLiteoverlayTile } from '@dlr-eoc/base-layers-raster';
 import { ICesiumControls, MapCesiumService } from '@dlr-eoc/map-cesium';
 import { MapOlService } from '@dlr-eoc/map-ol';
-import olMap from 'ol/Map';
 import { Cesium3DTileset, CesiumTerrainProvider, Credit, EllipsoidTerrainProvider, I3SDataProvider, createGooglePhotorealistic3DTileset } from '@cesium/engine';
 import testData from '@dlr-eoc/shared-assets/geojson/test.json';
 import { Feature } from 'ol';
@@ -37,7 +36,9 @@ export class RouteExampleCesiumComponent implements OnInit, OnDestroy {
     center: {
       lat: 40,
       lon: 13
-    }
+    },
+    rotation: 0,
+    viewAngle: 0
   };
 
   //Set inititial 2D/3D state
@@ -47,8 +48,8 @@ export class RouteExampleCesiumComponent implements OnInit, OnDestroy {
     @Inject('twoDlayerSvc') public twoDlayerSvc: LayersService,
     @Inject('threeDlayerSvc') public threeDlayerSvc: LayersService,
     public mapOlSvc: MapOlService,
-    public mapCesiumSvc: MapCesiumService,
-    public mapStateSvc: MapStateService
+    public mapStateSvc: MapStateService,
+    public mapCesiumSvc: MapCesiumService
   ) {
     this.controls = {
       infoBox: true,
@@ -274,17 +275,6 @@ export class RouteExampleCesiumComponent implements OnInit, OnDestroy {
   }
 
   async addTerrainlayer() {
-    // Terrain Datasource as Custom Layer
-    const bkgDGM5 = new CustomLayer({
-      name: 'BKG DGM5',
-      displayName: 'BKG DGM5',
-      id: 'bkg_dgm5',
-      custom_layer: await CesiumTerrainProvider.fromUrl('https://geoportal.de/3d/dgm5'),
-      visible: false,
-      attribution: `© <a href="https://www.bkg.de">Bundesamt für Kartographie und Geodäsie (BKG)</a>`,
-      description: 'BKG digital elevation modell with 5m grid resulution',
-      type: 'custom',
-    });
 
     // Dataservices from the masterportal: https://bitbucket.org/geowerkstatt-hamburg/masterportalapi/src/master/
     const hamburg_dgm = new CustomLayer({
@@ -313,7 +303,7 @@ export class RouteExampleCesiumComponent implements OnInit, OnDestroy {
     });
 
 
-    const layers = [bkgDGM5, hamburg_dgm, default_ellipsoid];
+    const layers = [hamburg_dgm, default_ellipsoid];
     layers.map(l => this.threeDlayerSvc.addLayer(l, 'Baselayers'));
   }
 
@@ -321,22 +311,29 @@ export class RouteExampleCesiumComponent implements OnInit, OnDestroy {
     if (this.is2dMap) {
       this.is2dMap = false;
     } else {
-      // If there is no camera roll, the flyTo is not necessary
-      if (this.mapCesiumSvc.viewer.camera.roll != 0) {
-        this.mapCesiumSvc.setNadirViewAngle();
-        // Wait for the cesim flyTo function to finish
-        setTimeout(() => {
-          //the openlayers map needs to be fully resetted, to avoid an error with the map controls
-          this.mapOlSvc.map = new olMap({});
+      this.mapCesiumSvc.setNadirViewAngle({
+        complete: () => {
           this.is2dMap = true;
-        }, 1500);
-      } else {
-        this.mapOlSvc.map = new olMap({});
-        this.is2dMap = true;
-      }
-
-
+        }
+      });
     }
+  }
+
+  setViewAngle() {
+    /** set map rotation with the MapStateService */
+    this.mapStateSvc.setViewAngle(45);
+  }
+  resetViewAngle() {
+    /** set map rotation with the MapStateService */
+    this.mapStateSvc.setViewAngle(0);
+  }
+  setRotation() {
+    /** set map rotation with the MapStateService */
+    this.mapStateSvc.setRotation(10);
+  }
+  resetRotation() {
+    /** set map rotation with the MapStateService, due to the rotation constraint small numbers are snapped to 0 */
+    this.mapStateSvc.setRotation(0);
   }
 }